Project fixes and optimization

+++ Assets/Plugins/TypedScenes/Core/TypedSceneStorage.cs
@@ -9,6 +9,10 @@ namespace IJunior.TypedScenes
         {
             var path = TypedSceneSettings.SavingDirectory + className + TypedSceneSettings.ClassExtension;
             Directory.CreateDirectory(TypedSceneSettings.SavingDirectory);
+
+            if (File.Exists(path) && File.ReadAllText(path) == sourceCode)
+                return;
+
             File.WriteAllText(path, sourceCode);
             AssetDatabase.ImportAsset(path, ImportAssetOptions.ForceUpdate);
         }
